метрики для конвертации размеров, завязанных на количество символов заливки, линии в элементах диаграмм
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@56133 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficeOdfFileW/source/OdfFormat/ods_table_state.h
@@ -84,7 +84,7 @@ public:
 		void set_table_dimension(int col, int row);
 
 	void add_column(office_element_ptr & elm, int repeated ,office_element_ptr & style);
-		void set_column_width(int width);
+		void set_column_width(double width);
 		void set_column_optimal_width(bool val);
 		void set_column_default_cell_style(std::wstring & style_name);
